RL: 2019-08-23
wsPrintForm : allow to open a  rpt report  instead of a  frx report.
WsSession : Create a sql select and setup for execution.
Add ffscontactemail form, it use for search email.
wsemailinfo : add on the three button search a method for open a form to search a email.
WsLogin : Allow user to enter lowercase or uppercase code and password in logon.
wsmuser2 : when search a user, show only the name.
mysql : add a replace for $ to Like when isfoxpro is false.